pip_requirements.txt
asgiref==3.7.2
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-appconf==1.0.6
django-browser-reload==1.12.1
django-compressor==4.4
djangorestframework==3.15.1
flatbuffers==24.3.25
idna==3.7
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
pandas==2.2.2
platformdirs==4.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
rcssmin==1.1.1
requests==2.32.3
requests-cache==1.2.0
retry-requests==2.0.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
